Overhauled scene graph & fixed a logic error
Scene graph now points to any arbitrary parent or child element.
The `transformof` method must be implemented for these elements.
This simplifies integrating the scene graph with custom constructions.

Fixed children of an element added to or removed from the scene graph
not triggering the respective events as well.

Added new unit test for the Transform World.